Fix Issue #311 - Complete support for all constants as Matlab function
Most constants are now usable through a matlab like call usign sizes.
This allows for easier initialization of matrices.

+++ modules/boost/simd/constant/include/boost/simd/constant/constants/eleven.hpp
@@ -57,7 +57,7 @@ namespace boost { namespace simd
      * \brief Define the tag Eleven of functor Eleven
      *        in namespace boost::simd::tag for toolbox boost.simd.constant
     **/
-    BOOST_SIMD_CONSTANT_REGISTER( Eleven, int, 11
+    BOOST_SIMD_CONSTANT_REGISTER( Eleven, double, 11
                                 ,0x41300000,0x4026000000000000ll
                                 );
   }
