+++ modules/boost/simd/constant/include/boost/simd/constant/constants/fifteen.hpp
@@ -57,7 +57,9 @@ namespace boost { namespace simd
      * \brief Define the tag Fifteen of functor Fifteen
      *        in namespace boost::simd::tag for toolbox boost.simd.constant
     **/
-    BOOST_SIMD_CONSTANT_REGISTER(Fifteen,int,15,0x41700000,0x402e000000000000ll);
+    BOOST_SIMD_CONSTANT_REGISTER( Fifteen,double
+                                , 15,0x41700000,0x402e000000000000ll
+                                );
   }
 
   BOOST_SIMD_CONSTANT_IMPLEMENTATION(boost::simd::tag::Fifteen, Fifteen)
